Improve tests (#662)
