+++ src/main/java/de/tracetronic/jenkins/plugins/ecutest/report/atx/installation/ATXSecretSetting.java
@@ -22,19 +22,18 @@ public class ATXSecretSetting extends ATXSetting<Secret> {
     private static final long serialVersionUID = 1L;
 
     /**
-     * Instantiates a new {@link ATXSecretSetting}.
+     * Instantiates a new {@link ATXSecretSetting} directly from a {@link Secret}.
      *
      * @param name  the name
      * @param group the settings group
      * @param value the current value
      */
-    //@DataBoundConstructor
     public ATXSecretSetting(final String name, final SettingsGroup group, final Secret value) {
         super(name, group, value);
     }
 
     /**
-     * Instantiates a new {@link ATXSecretSetting}.
+     * Instantiates a new {@link ATXSecretSetting} from a string value.
      *
      * @param name  the name
      * @param group the settings group
